refactor: relocate prior sync checksum module to common package, update import paths, and remove architecture diagram.
